🧹refractor(boundingbox): split generate method
