--- api/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "zk-maps-fileserver",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "express": "^4.17.2",
-    "shelljs": "^0.8.5"
-  }
-}
